+++ types/adapter/repositories/NodeLocalCommandRunner.d.ts
@@ -5,5 +5,6 @@ export declare class NodeLocalCommandRunner implements LocalCommandRunner {
         stderr: string;
         exitCode: number;
     }>;
+    spawnInteractive: (program: string, args: string[]) => void;
 }
 //# sourceMappingURL=NodeLocalCommandRunner.d.ts.map